+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix(response-interceptor): reduce responseInterceptor buffer churn
 - fix(ws): handle multi-server upgrade subscription and safe proxy shutdown
 - feat(router): add 'res' and 'options' to router function
+- feat(pathRewrite): add 'res' and 'options' to custom pathRewrite function
 
 ## [v4.0.0](https://github.com/chimurai/http-proxy-middleware/releases/tag/v4.0.0)